+++ src/optimagic/__init__.py
@@ -47,6 +47,7 @@ from optimagic.optimizers import pygad
 from optimagic.parameters.bounds import Bounds
 from optimagic.parameters.constraint_tools import check_constraints, count_free_params
 from optimagic.parameters.scaling import ScalingOptions
+from optimagic.typing import BatchEvaluator
 from optimagic.visualization.convergence_plot import convergence_plot
 from optimagic.visualization.history_plots import criterion_plot, params_plot
 from optimagic.visualization.profile_plot import profile_plot
@@ -85,6 +86,7 @@ __all__ = [
     "ScalingOptions",
     "MultistartOptions",
     "NumdiffOptions",
+    "BatchEvaluator",
     "FunctionValue",
     "LeastSquaresFunctionValue",
     "ScalarFunctionValue",